requirements.txt
# Python Research Repository for ISIC 2018 Dermoscopy
# Requirements file for master-thesis project

# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
scipy>=1.7.0

# Configuration and utilities
PyYAML>=5.4.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: Deep learning frameworks and tools
timm>=0.4.12  # PyTorch Image Models - for various architectures
albumentations>=1.0.0  # Advanced data augmentation

# Optional: Jupyter support for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Optional: Experiment tracking
# wandb>=0.12.0
# tensorboard>=2.6.0

# Optional: CAM methods (if using pytorch-grad-cam)
# pytorch-grad-cam>=1.3.0

# Development dependencies
pytest>=6.2.0
black>=21.7b0
flake8>=3.9.0
